fixed some errors resulted from merging previous PRs
